+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.10.0 (2026-05-13)
+
+
+### EZ Policy Enforcer
+
+* **enforcer:** Adjust sampling rate
+
+
+### Bug Fixes
+
+* streaming deadlock in Isolate Junction proxy
+
 ## 0.9.0 (2026-05-11)